requirements.txt
# API & Data
requests
pandas
numpy

# NLP & ML
scikit-learn
# PyTorch: for GPU support, install separately:
#   uv pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
torch
torchvision
torchaudio
transformers
datasets
tokenizers

# Visualization
matplotlib
seaborn
wordcloud

# Utilities
pyyaml
tqdm

# Testing
pytest

# Optional (for keyword extraction)
# keybert
